chore(parser): move Dictionary definition to openapi
