+++ src/openApi/common/interfaces/Dictionary.ts
