+++ src/openApi/index.ts
@@ -4,6 +4,8 @@ import { OpenApi } from './common/interfaces/OpenApi';
 import { parse as parseV2 } from './v2/index';
 import { parse as parseV3 } from './v3/index';
 
+export { OpenApi } from './common/interfaces/OpenApi';
+
 /**
  * Parse the OpenAPI specification to a Client model that contains
  * all the models, services and schema's we should output.
