+++ src/openApi/v2/interfaces/OpenApiHeader.ts
@@ -1,4 +1,4 @@
-import type { Dictionary } from '../../../types/generic';
+import type { Dictionary } from '../../common/interfaces/Dictionary';
 import type { OpenApiItems } from './OpenApiItems';
 
 /**
