+++ src/openApi/v3/interfaces/OpenApiLink.ts
@@ -1,4 +1,4 @@
-import type { Dictionary } from '../../../types/generic';
+import type { Dictionary } from '../../common/interfaces/Dictionary';
 import type { OpenApiReference } from './OpenApiReference';
 import type { OpenApiServer } from './OpenApiServer';
